pyproject.toml
[tool.poetry]
name = "crossmint-challenge"
version = "0.1.0"
description = "Megaverses are 2D spaces comprised of combinations of different astral objects: 🪐POLYanets with 🌙SOLoons around them and ☄comETHs floating around."
authors = ["Wolfgang <wleon@bakrypt.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
commitizen = "^3.30.0"
pre-commit = "^4.0.1"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
